Get building with msquic.hpp
