+++ src/perf/lib/PerfHelpers.h
@@ -16,7 +16,14 @@ Abstract:
 #endif
 
 #include <msquic.h>
+
+extern const QUIC_API_TABLE* MsQuic;
+
+#define QUIC_SKIP_GLOBAL_CONSTRUCTORS
+
 #include <quic_platform.h>
+#include <msquic.hpp>
+
 #include <msquichelper.h>
 #include <stdlib.h>
 #include <stdio.h>
@@ -27,7 +34,6 @@ Abstract:
 #include <new.h>
 #endif
 
-extern const QUIC_API_TABLE* MsQuic;
 #ifdef _KERNEL_MODE
 extern uint8_t SelfSignedSecurityHash[20];
 #else
@@ -36,143 +42,6 @@ extern QUIC_SEC_CONFIG_PARAMS* SelfSignedParams;
 extern bool IsSelfSignedValid;
 
 #define QUIC_TEST_SESSION_CLOSED    1
-
-struct QuicAddr {
-    QUIC_ADDR SockAddr;
-    QuicAddr() {
-        QuicZeroMemory(&SockAddr, sizeof(SockAddr));
-    }
-    QuicAddr(QUIC_ADDRESS_FAMILY af) {
-        QuicZeroMemory(&SockAddr, sizeof(SockAddr));
-        QuicAddrSetFamily(&SockAddr, af);
-    }
-    QuicAddr(QUIC_ADDRESS_FAMILY af, bool /*unused*/) {
-        QuicZeroMemory(&SockAddr, sizeof(SockAddr));
-        QuicAddrSetFamily(&SockAddr, af);
-        QuicAddrSetToLoopback(&SockAddr);
-    }
-    QuicAddr(const QuicAddr &Addr, uint16_t Port) {
-        SockAddr = Addr.SockAddr;
-        QuicAddrSetPort(&SockAddr, Port);
-    }
-    uint16_t GetPort() const { return QuicAddrGetPort(&SockAddr); }
-};
-
-template<typename T>
-class UniquePtr {
-public:
-    UniquePtr() noexcept = default;
-
-    explicit UniquePtr(T* _ptr) :
-        ptr{_ptr}
-    {
-    }
-
-    UniquePtr(const UniquePtr& other) = delete;
-    UniquePtr& operator=(const UniquePtr& other) = delete;
-
-    UniquePtr(UniquePtr&& other) noexcept {
-        this->ptr = other->ptr;
-        other->ptr = nullptr;
-    }
-
-    UniquePtr& operator=(UniquePtr&& other) noexcept {
-        if (this->ptr) {
-            delete this->ptr;
-        }
-        this->ptr = other->ptr;
-        other->ptr = nullptr;
-    }
-
-    ~UniquePtr() noexcept {
-        if (this->ptr) {
-            delete this->ptr;
-        }
-    }
-
-    void reset(T* lptr) noexcept {
-        if (this->ptr) {
-            delete this->ptr;
-        }
-        this->ptr = lptr;
-    }
-
-    T* release() noexcept {
-        T* tmp = ptr;
-        ptr = nullptr;
-        return tmp;
-    }
-
-    T* get() const noexcept { return ptr; }
-
-    T& operator*() const { return *ptr; }
-    T* operator->() const noexcept { return ptr; }
-    operator bool() const noexcept { return ptr != nullptr; }
-    bool operator == (T* _ptr) const noexcept { return ptr == _ptr; }
-    bool operator != (T* _ptr) const noexcept { return ptr != _ptr; }
-
-private:
-    T* ptr = nullptr;
-};
-
-template<typename T>
-class UniquePtr<T[]> {
-public:
-    UniquePtr() noexcept = default;
-
-    explicit UniquePtr(T* _ptr) :
-        ptr{_ptr}
-    {
-    }
-
-    UniquePtr(const UniquePtr& other) = delete;
-    UniquePtr& operator=(const UniquePtr& other) = delete;
-
-    UniquePtr(UniquePtr&& other) noexcept {
-        this->ptr = other->ptr;
-        other->ptr = nullptr;
-    }
-
-    UniquePtr& operator=(UniquePtr&& other) noexcept {
-        if (this->ptr) {
-            delete[] this->ptr;
-        }
-        this->ptr = other->ptr;
-        other->ptr = nullptr;
-    }
-
-    ~UniquePtr() noexcept {
-        if (this->ptr) {
-            delete[] this->ptr;
-        }
-    }
-
-    void reset(T* _ptr) noexcept {
-        if (this->ptr) {
-            delete[] this->ptr;
-        }
-        this->ptr = _ptr;
-    }
-
-    T* release() noexcept {
-        T* tmp = ptr;
-        ptr = nullptr;
-        return tmp;
-    }
-
-    T* get() const noexcept { return ptr; }
-
-    T& operator[](size_t i) const {
-        return *(ptr + i);
-    }
-
-    operator bool() const noexcept { return ptr != nullptr; }
-    bool operator == (T* _ptr) const noexcept { return ptr == _ptr; }
-    bool operator != (T* _ptr) const noexcept { return ptr != _ptr; }
-
-private:
-    T* ptr = nullptr;
-};
 
 inline
 int
@@ -192,177 +61,6 @@ WriteOutput(
     return 0;
 #endif
 }
-
-class MsQuicRegistration {
-    HQUIC Registration;
-public:
-    MsQuicRegistration() {
-        QuicZeroMemory(&Registration, sizeof(Registration));
-        if (QUIC_FAILED(MsQuic->RegistrationOpen(nullptr, &Registration))) {
-            Registration = nullptr;
-        }
-    }
-    ~MsQuicRegistration() {
-        if (Registration != nullptr) {
-            MsQuic->RegistrationClose(Registration);
-        }
-    }
-    bool IsValid() const { return Registration != nullptr; }
-    MsQuicRegistration(MsQuicRegistration& other) = delete;
-    MsQuicRegistration operator=(MsQuicRegistration& Other) = delete;
-    operator HQUIC () const {
-        return Registration;
-    }
-};
-
-struct MsQuicSession {
-    HQUIC Handle;
-    bool CloseAllConnectionsOnDelete;
-    MsQuicSession(_In_ const MsQuicRegistration& Reg, _In_z_ const char* RawAlpn)
-        : Handle(nullptr), CloseAllConnectionsOnDelete(false) {
-        if (!Reg.IsValid()) {
-            return;
-        }
-        QUIC_BUFFER Alpn;
-        Alpn.Buffer = (uint8_t*)RawAlpn;
-        Alpn.Length = (uint32_t)strlen(RawAlpn);
-        if (QUIC_FAILED(
-            MsQuic->SessionOpen(
-                Reg,
-                &Alpn,
-                1,
-                nullptr,
-                &Handle))) {
-            Handle = nullptr;
-        }
-    }
-    ~MsQuicSession() {
-        if (Handle != nullptr) {
-            if (CloseAllConnectionsOnDelete) {
-                MsQuic->SessionShutdown(
-                    Handle,
-                    QUIC_CONNECTION_SHUTDOWN_FLAG_SILENT,
-                    QUIC_TEST_SESSION_CLOSED);
-            }
-            MsQuic->SessionClose(Handle);
-        }
-    }
-    bool IsValid() const {
-        return Handle != nullptr;
-    }
-    MsQuicSession(MsQuicSession& other) = delete;
-    MsQuicSession operator=(MsQuicSession& Other) = delete;
-    operator HQUIC () const {
-        return Handle;
-    }
-    void SetAutoCleanup() {
-        CloseAllConnectionsOnDelete = true;
-    }
-    void Shutdown(
-        _In_ QUIC_CONNECTION_SHUTDOWN_FLAGS Flags,
-        _In_ QUIC_UINT62 ErrorCode
-        ) {
-        MsQuic->SessionShutdown(Handle, Flags, ErrorCode);
-    }
-    QUIC_STATUS
-    SetTlsTicketKey(
-        _In_reads_bytes_(44)
-            const uint8_t* const Buffer
-        ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_TLS_TICKET_KEY,
-                44,
-                Buffer);
-    }
-    QUIC_STATUS
-    SetPeerBidiStreamCount(
-        uint16_t value
-        ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_PEER_BIDI_STREAM_COUNT,
-                sizeof(value),
-                &value);
-    }
-    QUIC_STATUS
-    SetPeerUnidiStreamCount(
-        uint16_t value
-        ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_PEER_UNIDI_STREAM_COUNT,
-                sizeof(value),
-                &value);
-    }
-    QUIC_STATUS
-    SetIdleTimeout(
-        uint64_t value  // milliseconds
-        ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_IDLE_TIMEOUT,
-                sizeof(value),
-                &value);
-    }
-    QUIC_STATUS
-    SetDisconnectTimeout(
-        uint32_t value  // milliseconds
-        ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_DISCONNECT_TIMEOUT,
-                sizeof(value),
-                &value);
-    }
-    QUIC_STATUS
-    SetMaxBytesPerKey(
-        uint64_t value
-        ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_MAX_BYTES_PER_KEY,
-                sizeof(value),
-                &value);
-    }
-    QUIC_STATUS
-    SetDatagramReceiveEnabled(
-        bool value
-        ) {
-        BOOLEAN Value = value ? TRUE : FALSE;
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_DATAGRAM_RECEIVE_ENABLED,
-                sizeof(Value),
-                &Value);
-    }
-    QUIC_STATUS
-    SetServerResumptionLevel(
-        QUIC_SERVER_RESUMPTION_LEVEL Level
-    ) {
-        return
-            MsQuic->SetParam(
-                Handle,
-                QUIC_PARAM_LEVEL_SESSION,
-                QUIC_PARAM_SESSION_SERVER_RESUMPTION_LEVEL,
-                sizeof(Level),
-                &Level);
-    }
-};
 
 struct MsQuicListener {
     HQUIC Handle{ nullptr };
@@ -481,51 +179,6 @@ struct MsQuicSecurityConfig {
 
     QUIC_SEC_CONFIG* SecurityConfig {nullptr};
 };
-
-struct ListenerScope {
-    HQUIC Handle;
-    ListenerScope() : Handle(nullptr) { }
-    ListenerScope(HQUIC handle) : Handle(handle) { }
-    ~ListenerScope() { if (Handle) { MsQuic->ListenerClose(Handle); } }
-    operator HQUIC() const { return Handle; }
-};
-
-struct ConnectionScope {
-    HQUIC Handle;
-    ConnectionScope() : Handle(nullptr) { }
-    ConnectionScope(HQUIC handle) : Handle(handle) { }
-    ~ConnectionScope() { if (Handle) { MsQuic->ConnectionClose(Handle); } }
-    operator HQUIC() const { return Handle; }
-};
-
-struct StreamScope {
-    HQUIC Handle;
-    StreamScope() : Handle(nullptr) { }
-    StreamScope(HQUIC handle) : Handle(handle) { }
-    ~StreamScope() { if (Handle) { MsQuic->StreamClose(Handle); } }
-    operator HQUIC() const { return Handle; }
-};
-
-struct EventScope {
-    QUIC_EVENT Handle;
-    EventScope() { QuicEventInitialize(&Handle, FALSE, FALSE); }
-    EventScope(QUIC_EVENT event) : Handle(event) { }
-    ~EventScope() { QuicEventUninitialize(Handle); }
-    operator QUIC_EVENT() const { return Handle; }
-};
-
-struct QuicBufferScope {
-    QUIC_BUFFER* Buffer;
-    QuicBufferScope() : Buffer(nullptr) { }
-    QuicBufferScope(uint32_t Size) : Buffer((QUIC_BUFFER*) new uint8_t[sizeof(QUIC_BUFFER) + Size]) {
-        QuicZeroMemory(Buffer, sizeof(*Buffer) + Size);
-        Buffer->Length = Size;
-        Buffer->Buffer = (uint8_t*)(Buffer + 1);
-    }
-    operator QUIC_BUFFER* () { return Buffer; }
-    ~QuicBufferScope() { if (Buffer) { delete[](uint8_t*) Buffer; } }
-};
-
 
 struct CountHelper {
     long RefCount;
